requirements.txt
pandas
gspread
google-auth
typer
rich
pytest
